P4ADEV-2412 moved class
