+++ src/main/java/it/gov/pagopa/payhub/activities/service/exportflow/debtposition/BaseExportFileService.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.payhub.activities.service.exportflow.debtposition;
 
-import it.gov.pagopa.payhub.activities.dto.export.debtposition.ExportFileResult;
+import it.gov.pagopa.payhub.activities.dto.exportflow.ExportFileResult;
 import it.gov.pagopa.payhub.activities.exception.exportflow.ExportFileNotFoundException;
 import it.gov.pagopa.payhub.activities.exception.exportflow.InvalidExportStatusException;
 import it.gov.pagopa.payhub.activities.service.CsvService;
